Move unit tests into existing private-actions suite
